backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.1
torch==2.0.1
torchvision==0.15.2
Werkzeug==2.3.7
